Refactor staker transitions
